requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
tensorflow==2.13.0
numpy==1.24.3
pillow==10.0.0
opencv-python-headless==4.8.1.78
opencv-python-headless==4.8.1.78
python-dotenv==1.0.0
pydantic==2.5.0
fastapi-cli==0.0.1
scikit-learn==1.3.0
